fix 버그 수정(24.07.04)
