+++ app/src/main/java/com/nomorelateness/donotlate/feature/main/presentation/view/MainFragment.kt
@@ -62,7 +62,6 @@ class MainFragment : Fragment() {
 
         observeViewModel()
         initButton()
-        initDarMode()
         setTextUserName()
     }
 
@@ -97,20 +96,6 @@ class MainFragment : Fragment() {
         startMiniGame()
         startMyPromise()
         startFriends()
-    }
-
-    private fun initDarMode() {
-        val sharedPref = activity?.getPreferences(Context.MODE_PRIVATE) ?: return
-
-        val sharedPrefValue = resources.getString(R.string.preference_darkMode_key)
-        val darkModeValue =
-            sharedPref.getString(getString(R.string.preference_darkMode_key), sharedPrefValue)
-
-        if (darkModeValue == "darkModeOn") {
-            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_YES)
-        } else {
-            AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-        }
     }
 
     private fun startPlace() {
